requirements.txt
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
python-dateutil>=2.8.2
pyyaml>=6.0.1
tabulate>=0.9.0
matplotlib>=3.7.0
pandas>=2.0.0

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0

# Reporting & Export
reportlab>=4.0.0
openpyxl>=3.1.0

# Statistics (Enhanced)
scipy>=1.11.0
statsmodels>=0.14.0
numpy>=1.24.0

# Interactive Visualizations
plotly>=5.17.0
kaleido>=0.2.1  # For static image export from Plotly

# AI Insights
openai>=1.0.0
python-dotenv>=1.0.0
